Addressing post-review feedback on r235619. https://bugs.webkit.org/show_bug.cgi?id=187925
Unreviewed.

* testing/Internals.cpp:
(WebCore::Internals::primaryScreenDisplayID):
* testing/Internals.h:



Canonical link: https://commits.webkit.org/204499@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@235916 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/testing/Internals.cpp
@@ -4715,7 +4715,7 @@ void Internals::notifyResourceLoadObserver()
     ResourceLoadObserver::shared().notifyObserver();
 }
 
-unsigned long Internals::primaryScreenDisplayID()
+unsigned Internals::primaryScreenDisplayID()
 {
 #if PLATFORM(MAC)
     return WebCore::primaryScreenDisplayID();
